`load_paths` option
Use this to specify custom load paths.  These will be used with
`@import` statements.

The engine's load paths will always include the 'source_path' option
provided by Dassets when the engine is configured.

Note: this also adds a test to ensure that sass is compiled with
the appropriate options.  This required updating to the latest
Assert to get its stubbing API.